fix(engine): widen intermediate product to i128 in demo claim linkage rate calculation
Prevent i64 overflow in linked_claims_rate_millionths() when
linked_claims is large, using the same i128 widening pattern
applied to interleaving_explorer and translation_validation_receipt.

+++ crates/franken-engine/src/demo_claim_linkage_gate.rs
@@ -291,7 +291,7 @@ impl LinkageGateDecision {
         if self.total_claims == 0 {
             return 0;
         }
-        (self.linked_claims as i64 * MILLION) / self.total_claims as i64
+        (self.linked_claims as i128 * MILLION as i128 / self.total_claims as i128) as i64
     }
 }
